requirements.txt
actionlib==1.12.1
angles==1.9.12
ansi2html==1.8.0
asttokens==2.4.1
attrs==23.1.0
backcall==0.2.0
bagpy==0.5
bitarray==2.8.2
bitstring==4.1.2
bondpy==1.8.5
camera_calibration==1.15.2
camera_calibration_parsers==1.11.13
catkin==0.7.29
catkin-pkg @ file:///home/conda/feedstock_root/build_artifacts/catkin_pkg_1694651879177/work
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
comm==0.2.1
ConfigArgParse==1.7
contourpy==1.1.1
controller_manager==0.18.4
controller_manager_msgs==0.18.4
cv_bridge==1.13.1
cycler==0.12.1
dash==2.14.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
diagnostic_analysis==1.9.7
diagnostic_common_diagnostics==1.9.7
diagnostic_updater==1.9.7
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1675116244235/work
docutils @ file:///home/conda/feedstock_root/build_artifacts/docutils_1695300728982/work
dynamic_reconfigure==1.6.5
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
Flask==2.2.5
fonttools==4.43.1
gazebo_plugins==2.8.7
gazebo_ros==2.8.7
gencpp==0.6.6
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.17
gennodejs==2.0.1
genpy==0.6.16
gnupg==2.3.1
haversine==2.8.0
idna==3.4
image_geometry==1.13.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
interactive-markers==1.11.5
ipyfilechooser==0.6.0
ipython==8.16.1
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.0.3
joint_state_publisher==1.12.15
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_core==5.4.0
jupyterlab-widgets==3.0.9
kdl-parser-py==1.13.3
kiwisolver==1.4.5
laser_geometry==1.6.7
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
message_filters==1.14.13
motmetrics==1.4.0
motrackers==0.0.2
nbformat==5.7.0
nest-asyncio==1.5.8
numpy==1.22.4
open3d==0.17.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.1.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.18.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py3rosmsgs==1.18.2
pycryptodomex==3.19.0
Pygments==2.16.1
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1690737849915/work
pyserial @ file:///tmp/build/80754af9/pyserial_1594419199743/work
pytest==7.4.3
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python_qt_binding==0.4.4
pytz==2023.3.post1
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1695373594639/work
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
referencing==0.30.2
requests==2.31.0
resource_retriever==1.12.7
retrying==1.3.4
ros_numpy==0.0.3
rosbag==1.14.13
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.13
roslaunch==1.14.13
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
rosmsg==1.14.13
rosnode==1.14.13
rosparam==1.14.13
rospkg @ file:///home/conda/feedstock_root/build_artifacts/rospkg_1679367642027/work
rospy==1.14.13
rosservice==1.14.13
rostest==1.14.13
rostopic==1.14.13
rosunit==1.14.9
roswtf==1.14.13
rpds-py==0.10.6
rqt-image-view==0.4.17
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.9
rqt_dep==0.4.9
rqt_graph==0.4.11
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.8
rqt_logger_level==0.4.8
rqt_msg==0.4.8
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.8
rqt_publisher==0.4.8
rqt_py_common==0.5.3
rqt_py_console==0.4.8
rqt_robot_dashboard==0.5.7
rqt_robot_steering==0.5.10
rqt_runtime_monitor==0.5.7
rqt_service_caller==0.4.8
rqt_shell==0.4.9
rqt_srv==0.4.8
rqt_tf_tree==0.6.0
rqt_top==0.4.8
rqt_topic==0.4.11
rqt_web==0.4.8
rviz==1.13.30
scipy==1.11.4
seaborn==0.13.0
sensor-msgs==1.12.8
setuptools-scm==8.0.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smach==2.0.3
smach_ros==2.0.3
smclib==1.8.5
stack-data==0.6.3
tenacity==8.2.3
tf==1.12.1
tf2_geometry_msgs==0.6.5
tf2_kdl==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf_conversions==1.12.1
tomli==2.0.1
topic_tools==1.14.13
traitlets==5.12.0
typing_extensions==4.8.0
tzdata==2023.3
urdfdom-py==0.4.6
urllib3==2.0.7
wcwidth==0.2.8
Werkzeug==2.2.3
widgetsnbextension==4.0.9
xacro==1.13.20
xmltodict==0.13.0
zipp==3.17.0
